requirements.txt
ccxt==2.7.17
pandas==1.5.3
numpy==1.24.2
ta==0.10.1
python-binance==1.0.15
